Split files.
Probably make actor classes public, but instantiation private and add an instantiator class in FieldManager.
