+++ MapleServer2/Managers/FieldManager.cs
@@ -2,7 +2,6 @@
 using Maple2.Trigger;
 using Maple2.Trigger.Enum;
 using Maple2Storage.Enums;
-using Maple2Storage.Tools;
 using Maple2Storage.Types;
 using Maple2Storage.Types.Metadata;
 using MaplePacketLib2.Tools;
@@ -21,7 +20,7 @@ namespace MapleServer2.Managers;
 // TODO: This needs to be thread safe
 // TODO: FieldManager probably needs its own thread to send updates about user position
 // This seems to be done every ~2s rather than on every update.
-public class FieldManager
+public partial class FieldManager
 {
     private static readonly TriggerLoader TriggerLoader = new();
 
@@ -367,15 +366,7 @@ public class FieldManager
             return npc;
         }
 
-        Mob mob = WrapMob(npcId);
-        mob.Coord = coord;
-        mob.Rotation = rotation;
-        if (animation != default)
-        {
-            mob.Animation = animation;
-        }
-        AddMob(mob);
-        return mob;
+        return RequestMob(npcId, coord, rotation, animation);
     }
 
     public IFieldActor<NpcMetadata> RequestMob(int mobId, CoordF coord = default, CoordF rotation = default, short animation = default)
@@ -772,575 +763,6 @@ public class FieldManager
         return new(objectId, mobId);
     }
 
-    // This class is private to ensure that callers must first request entry.
-    private class FieldObject<T> : IFieldObject<T>
-    {
-        public int ObjectId { get; set; }
-        public T Value { get; }
-
-        public virtual CoordF Coord { get; set; }
-        public CoordF Rotation { get; set; }
-        public short LookDirection
-        {
-            get => (short) (Rotation.Z * 10);
-            set => Rotation = CoordF.From(Rotation.X, Rotation.Y, value / 10);
-        }
-
-        public FieldObject(int objectId, T value)
-        {
-            ObjectId = objectId;
-            Value = value;
-        }
-    }
-
-    private abstract class FieldActor<T> : FieldObject<T>, IFieldActor<T>
-    {
-        public CoordF Velocity { get; set; }
-        public short Animation { get; set; }
-
-        public virtual Stats Stats { get; set; }
-        public bool IsDead { get; set; }
-
-        public List<Status> Statuses { get; set; }
-        public SkillCast SkillCast { get; set; }
-        public bool OnCooldown { get; set; }
-
-        public FieldActor(int objectId, T value) : base(objectId, value) { }
-
-        public virtual void UpdateFixed() { }
-
-        public virtual void Update() { }
-
-        public virtual void MoveBy(CoordF displacement)
-        {
-            // TODO: Break displacement into segments, then set path.
-            Velocity = displacement;
-        }
-
-        public virtual void MoveTo(CoordF target)
-        {
-            // TODO: Perform pathfinding, then set path.
-            Velocity = target - Coord;
-        }
-
-        public virtual void Cast(SkillCast skillCast)
-        {
-            SkillCast = skillCast;
-
-            // TODO: Move this and all others combat cases like recover sp to its own class.
-            // Since the cast is always sent by the skill, we have to check buffs even when not doing damage.
-            if (skillCast.IsBuffToOwner() || skillCast.IsBuffToEntity() || skillCast.IsBuffShield() || skillCast.IsDebuffToOwner())
-            {
-                Status status = new(skillCast, ObjectId, ObjectId, 1);
-                //StatusHandler.Handle(Value.Session, status);
-            }
-        }
-
-        public virtual void RecoverHp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Hp];
-                if (stat.Total < stat.Bonus)
-                {
-                    stat.Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                }
-            }
-        }
-
-        public virtual void ConsumeHp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Hp];
-                stat.Decrease(Math.Min(amount, stat.Total));
-            }
-        }
-
-        public virtual void RecoverSp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Spirit];
-                if (stat.Total < stat.Bonus)
-                {
-                    stat.Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                }
-            }
-        }
-
-        public virtual void ConsumeSp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Spirit];
-                Stats[StatId.Spirit].Decrease(Math.Min(amount, stat.Total));
-            }
-        }
-
-        public virtual void RecoverStamina(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Stamina];
-                if (stat.Total < stat.Bonus)
-                {
-                    Stats[StatId.Stamina].Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                }
-            }
-        }
-
-        public virtual void ConsumeStamina(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Stamina];
-                Stats[StatId.Stamina].Decrease(Math.Min(amount, stat.Total));
-            }
-        }
-
-        public virtual void Damage(DamageHandler damage)
-        {
-            Stat health = Stats[StatId.Hp];
-            health.Decrease((long) damage.Damage);
-            if (health.Total <= 0)
-            {
-                Perish();
-            }
-        }
-
-        public virtual void Perish()
-        {
-            IsDead = true;
-        }
-    }
-
-    private class Character : FieldActor<Player>
-    {
-        public override CoordF Coord
-        {
-            get => Value.Coord;
-            set => Value.Coord = value;
-        }
-
-        public override Stats Stats
-        {
-            get => Value.Stats;
-            set => Value.Stats = value;
-        }
-
-        private CancellationTokenSource CombatCTS;
-
-        private Task HpRegenThread;
-        private Task SpRegenThread;
-        private Task StaRegenThread;
-
-        public Character(int objectId, Player value) : base(objectId, value)
-        {
-            if (HpRegenThread == null || HpRegenThread.IsCompleted)
-            {
-                HpRegenThread = StartRegen(StatId.Hp, StatId.HpRegen, StatId.HpRegenInterval);
-            }
-            if (SpRegenThread == null || SpRegenThread.IsCompleted)
-            {
-                SpRegenThread = StartRegen(StatId.Spirit, StatId.SpRegen, StatId.SpRegenInterval);
-            }
-            if (StaRegenThread == null || StaRegenThread.IsCompleted)
-            {
-                StaRegenThread = StartRegen(StatId.Stamina, StatId.StaminaRegen, StatId.StaminaRegenInterval);
-            }
-        }
-
-        public override void Cast(SkillCast skillCast)
-        {
-            int spiritCost = skillCast.GetSpCost();
-            int staminaCost = skillCast.GetStaCost();
-
-            if (Value.Stats[StatId.Spirit].Total >= spiritCost && Value.Stats[StatId.Stamina].Total >= staminaCost)
-            {
-                SkillCast = skillCast;
-
-                ConsumeSp(spiritCost);
-                ConsumeStamina(staminaCost);
-                Value.Session.SendNotice(skillCast.SkillId.ToString());
-
-                // TODO: Move this and all others combat cases like recover sp to its own class.
-                // Since the cast is always sent by the skill, we have to check buffs even when not doing damage.
-                if (skillCast.IsBuffToOwner() || skillCast.IsBuffToEntity() || skillCast.IsBuffShield() || skillCast.IsDebuffToOwner())
-                {
-                    Status status = new(skillCast, ObjectId, ObjectId, 1);
-                    StatusHandler.Handle(Value.Session, status);
-                }
-
-                StartCombatStance();
-            }
-        }
-
-        public override void RecoverHp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Hp];
-                if (stat.Total < stat.Bonus)
-                {
-                    stat.Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                    Value.Session.Send(StatPacket.UpdateStats(this, StatId.Hp));
-                }
-            }
-        }
-
-        public override void ConsumeHp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Hp];
-                stat.Decrease(Math.Min(amount, stat.Total));
-            }
-
-            if (HpRegenThread == null || HpRegenThread.IsCompleted)
-            {
-                HpRegenThread = StartRegen(StatId.Hp, StatId.HpRegen, StatId.HpRegenInterval);
-            }
-        }
-
-        public override void RecoverSp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Spirit];
-                if (stat.Total < stat.Bonus)
-                {
-                    stat.Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                    Value.Session.Send(StatPacket.UpdateStats(this, StatId.Spirit));
-                }
-            }
-        }
-
-        public override void ConsumeSp(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Spirit];
-                Stats[StatId.Spirit].Decrease(Math.Min(amount, stat.Total));
-            }
-
-            if (SpRegenThread == null || SpRegenThread.IsCompleted)
-            {
-                SpRegenThread = StartRegen(StatId.Spirit, StatId.SpRegen, StatId.SpRegenInterval);
-            }
-        }
-
-        public override void RecoverStamina(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Stamina];
-                if (stat.Total < stat.Bonus)
-                {
-                    Stats[StatId.Stamina].Increase(Math.Min(amount, stat.Bonus - stat.Total));
-                    Value.Session.Send(StatPacket.UpdateStats(this, StatId.Stamina));
-                }
-            }
-        }
-
-        public override void ConsumeStamina(int amount)
-        {
-            if (amount <= 0)
-            {
-                return;
-            }
-
-            lock (Stats)
-            {
-                Stat stat = Stats[StatId.Stamina];
-                Stats[StatId.Stamina].Decrease(Math.Min(amount, stat.Total));
-            }
-
-            if (StaRegenThread == null || StaRegenThread.IsCompleted)
-            {
-                StaRegenThread = StartRegen(StatId.Stamina, StatId.StaminaRegen, StatId.StaminaRegenInterval);
-            }
-        }
-
-        private Task StartRegen(StatId statId, StatId regenStatId, StatId timeStatId)
-        {
-            // TODO: merge regen updates with larger packets
-            return Task.Run(async () =>
-            {
-                while (true)
-                {
-                    await Task.Delay(Stats[timeStatId].Total);
-
-                    lock (Stats)
-                    {
-                        if (Stats[statId].Total >= Stats[statId].Bonus)
-                        {
-                            return;
-                        }
-
-                        // TODO: Check if regen-enabled
-                        AddStatRegen(statId, regenStatId);
-                        Value.Session?.FieldManager.BroadcastPacket(StatPacket.UpdateStats(this, statId));
-                        if (Value.Party != null)
-                        {
-                            Value.Party.BroadcastPacketParty(PartyPacket.UpdateHitpoints(Value));
-                        }
-                    }
-                }
-            });
-        }
-
-        public Task StartCombatStance()
-        {
-            // Refresh out-of-combat timer
-            CombatCTS?.Cancel();
-            CancellationTokenSource cts = new();
-            cts.Token.Register(() => cts.Dispose());
-            CombatCTS = cts;
-
-            // Enter combat
-            Value.Session.FieldManager.BroadcastPacket(UserBattlePacket.UserBattle(this, true));
-            return Task.Run(async () =>
-            {
-                await Task.Delay(5000);
-
-                if (!cts.Token.IsCancellationRequested)
-                {
-                    CombatCTS = null;
-                    cts.Dispose();
-                    Value.Session?.FieldManager.BroadcastPacket(UserBattlePacket.UserBattle(this, false));
-                }
-            }, cts.Token);
-        }
-
-        private void AddStatRegen(StatId statIndex, StatId regenStatIndex)
-        {
-            int regenAmount = Stats[regenStatIndex].Total;
-            if (regenAmount <= 0)
-            {
-                return;
-            }
-
-            Stat stat = Stats[statIndex];
-            lock (stat)
-            {
-                if (stat.Total < stat.Bonus)
-                {
-                    int missingAmount = stat.Bonus - stat.Total;
-                    stat.Increase(Math.Clamp(regenAmount, 0, missingAmount));
-                }
-            }
-        }
-    }
-
-    private class Npc : FieldActor<NpcMetadata>, INpc
-    {
-        public NpcState State { get; set; }
-        public NpcAction Action { get; set; }
-        public MobMovement Movement { get; set; }
-
-        public Npc(int objectId, int npcId) : this(objectId, NpcMetadataStorage.GetNpcMetadata(npcId)) { }
-
-        public Npc(int objectId, NpcMetadata metadata) : base(objectId, metadata)
-        {
-            Animation = 255;
-            Stats = new(metadata);
-            State = NpcState.Normal;
-            Action = NpcAction.Idle;
-        }
-    }
-
-    private class Mob : FieldActor<NpcMetadata>, INpc
-    {
-        private readonly MobAI AI;
-        public IFieldObject<MobSpawn> OriginSpawn;
-
-        private CoordF SpawnDistance;
-
-        public NpcState State { get; set; }
-        public NpcAction Action { get; set; }
-        public MobMovement Movement { get; set; }
-        public IFieldActor<Player> Target;
-
-        public Mob(int objectId, int mobId) : this(objectId, NpcMetadataStorage.GetNpcMetadata(mobId)) { }
-
-        public Mob(int objectId, NpcMetadata metadata) : base(objectId, metadata)
-        {
-            Animation = AnimationStorage.GetSequenceIdBySequenceName(metadata.Model, "Idle_A");
-            AI = MobAIManager.GetAI(metadata.AiInfo);
-            Stats = new(metadata);
-            State = NpcState.Normal;
-        }
-
-        public void Attack()
-        {
-            int roll = RandomProvider.Get().Next(100);
-            for (int i = 0; i < Value.NpcMetadataSkill.SkillIds.Length; i++)
-            {
-                if (roll < Value.NpcMetadataSkill.SkillProbs[i])
-                {
-                    // Rolled this skill.
-                    Cast(new SkillCast(Value.NpcMetadataSkill.SkillIds[i], Value.NpcMetadataSkill.SkillLevels[i]));
-                    StartSkillTimer((Value.NpcMetadataSkill.SkillCooldown > 0) ? Value.NpcMetadataSkill.SkillCooldown : 1000);
-                }
-
-                roll -= Value.NpcMetadataSkill.SkillProbs[i];
-            }
-        }
-
-        private Task StartSkillTimer(int CooldownMilliseconds)
-        {
-            return Task.Run(async () =>
-            {
-                await Task.Delay(CooldownMilliseconds);
-
-                OnCooldown = false;
-            });
-        }
-
-        public void Act()
-        {
-            if (AI == null)
-            {
-                return;
-            }
-
-            (string actionName, NpcAction actionType) = AI.GetAction(this);
-
-            if (actionName != null)
-            {
-                Animation = AnimationStorage.GetSequenceIdBySequenceName(Value.Model, actionName);
-            }
-            Action = actionType;
-            Movement = AI.GetMovementAction(this);
-
-            switch (Action)
-            {
-                case NpcAction.Idle:
-                case NpcAction.Bore:
-                    Move(MobMovement.Hold); // temp, maybe remove the option to specify movement in AI
-                    break;
-                case NpcAction.Walk:
-                case NpcAction.Run:
-                    Move(Movement);
-                    break;
-                case NpcAction.Skill:
-                    // Cast skill
-                    if (!OnCooldown)
-                    {
-                        Attack();
-                        Move(MobMovement.Hold);
-                        break;
-                    }
-
-                    Move(Movement);
-                    break;
-                case NpcAction.Jump:
-                default:
-                    break;
-            }
-        }
-
-        public void Move(MobMovement moveType)
-        {
-            Random rand = RandomProvider.Get();
-
-            switch (moveType)
-            {
-                case MobMovement.Patrol:
-                    // Fallback Dummy Movement
-                    int moveDistance = rand.Next(0, Value.MoveRange);
-                    short moveDir = (short) rand.Next(-1800, 1800);
-
-                    Velocity = CoordF.From(moveDistance, moveDir);
-                    // Keep near spawn
-                    if ((SpawnDistance - Velocity).Length() >= Block.BLOCK_SIZE * 2)
-                    {
-                        moveDir = (short) SpawnDistance.XYAngle();
-                        Velocity = CoordF.From(Block.BLOCK_SIZE, moveDir);
-                    }
-
-                    LookDirection = moveDir; // looking direction of the monster
-                    break;
-                case MobMovement.Follow: // move towards target
-                    Velocity = CoordF.From(0, 0, 0);
-                    break;
-                case MobMovement.Strafe: // move around target
-                case MobMovement.Run: // move away from target
-                case MobMovement.LookAt:
-                case MobMovement.Hold:
-                default:
-                    Velocity = CoordF.From(0, 0, 0);
-                    break;
-            }
-
-            SpawnDistance -= Velocity;
-        }
-
-        public override void Perish()
-        {
-            IsDead = true;
-            State = NpcState.Dead;
-            int randAnim = RandomProvider.Get().Next(Value.StateActions[NpcState.Dead].Length);
-            Animation = AnimationStorage.GetSequenceIdBySequenceName(Value.Model, Value.StateActions[NpcState.Dead][randAnim].Item1);
-        }
-    }
-
     private Task StartMapLoop()
     {
         return Task.Run(async () =>
@@ -1499,4 +921,31 @@ public class FieldManager
     {
         SkipScene = enable;
     }
+
+    // This class is private to ensure that callers must first request entry.
+    private class FieldObject<T> : IFieldObject<T>
+    {
+        public int ObjectId { get; set; }
+        public T Value { get; }
+
+        public virtual CoordF Coord { get; set; }
+        public CoordF Rotation { get; set; }
+        public short LookDirection
+        {
+            get => (short) (Rotation.Z * 10);
+            set => Rotation = CoordF.From(Rotation.X, Rotation.Y, value / 10);
+        }
+
+        public FieldObject(int objectId, T value)
+        {
+            ObjectId = objectId;
+            Value = value;
+        }
+    }
+
+    private abstract partial class FieldActor<T> { }
+    private partial class Character { }
+    private partial class Mob { }
+    private partial class Npc { }
+    private abstract partial class FieldActor<T> { }
 }
